requirements.txt
streamlit>=1.30.0
python-dotenv>=1.0.0
PyPDF2>=3.0.0
langchain>=0.1.0
langchain-google-genai>=0.0.5
faiss-cpu>=1.7.4
google-generativeai>=0.3.0
requests>=2.31.0